length is int

+++ src/FilmForumModels/Dtos/EpisodeDtos/GetDetailedEpisodeDto.cs
@@ -33,7 +33,7 @@ public class GetDetailedEpisodeDto
 
     public int SeasonNumber { get; set; } = 0;
 
-    public float Length { get; set; } = 0.0f;
+    public int Length { get; set; } = 0;
 
     public int Year { get; set; } = 2023;
